+++ src/main/java/com/algorand/algosdk/v2/client/algod/GetPendingTransactions.java
@@ -28,9 +28,29 @@ public class GetPendingTransactions extends Query {
         return this;
     }
 
+   /**
+    * Execute the query.
+    * @return the query response object.
+    * @throws Exception
+    */
     @Override
     public Response<PendingTransactionsResponse> execute() throws Exception {
         Response<PendingTransactionsResponse> resp = baseExecute();
+        resp.setValueType(PendingTransactionsResponse.class);
+        return resp;
+    }
+
+   /**
+    * Execute the query with custom headers, there must be an equal number of keys and values
+    * or else an error will be generated.
+    * @param headers an array of header keys
+    * @param values an array of header values
+    * @return the query response object.
+    * @throws Exception
+    */
+    @Override
+    public Response<PendingTransactionsResponse> execute(String[] headers, String[] values) throws Exception {
+        Response<PendingTransactionsResponse> resp = baseExecute(headers, values);
         resp.setValueType(PendingTransactionsResponse.class);
         return resp;
     }