+++ src/main/java/com/algorand/algosdk/v2/client/algod/GetSupply.java
@@ -18,9 +18,29 @@ public class GetSupply extends Query {
         super(client, new HttpMethod("get"));
     }
 
+   /**
+    * Execute the query.
+    * @return the query response object.
+    * @throws Exception
+    */
     @Override
     public Response<SupplyResponse> execute() throws Exception {
         Response<SupplyResponse> resp = baseExecute();
+        resp.setValueType(SupplyResponse.class);
+        return resp;
+    }
+
+   /**
+    * Execute the query with custom headers, there must be an equal number of keys and values
+    * or else an error will be generated.
+    * @param headers an array of header keys
+    * @param values an array of header values
+    * @return the query response object.
+    * @throws Exception
+    */
+    @Override
+    public Response<SupplyResponse> execute(String[] headers, String[] values) throws Exception {
+        Response<SupplyResponse> resp = baseExecute(headers, values);
         resp.setValueType(SupplyResponse.class);
         return resp;
     }