+++ src/main/java/com/algorand/algosdk/v2/client/indexer/LookupBlock.java
@@ -24,9 +24,29 @@ public class LookupBlock extends Query {
         this.roundNumber = roundNumber;
     }
 
+   /**
+    * Execute the query.
+    * @return the query response object.
+    * @throws Exception
+    */
     @Override
     public Response<Block> execute() throws Exception {
         Response<Block> resp = baseExecute();
+        resp.setValueType(Block.class);
+        return resp;
+    }
+
+   /**
+    * Execute the query with custom headers, there must be an equal number of keys and values
+    * or else an error will be generated.
+    * @param headers an array of header keys
+    * @param values an array of header values
+    * @return the query response object.
+    * @throws Exception
+    */
+    @Override
+    public Response<Block> execute(String[] headers, String[] values) throws Exception {
+        Response<Block> resp = baseExecute(headers, values);
         resp.setValueType(Block.class);
         return resp;
     }