+++ src/test/java/com/algorand/algosdk/unit/RunCucumberUnitTest.java
@@ -5,6 +5,6 @@ import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(plugin = {"progress"}, tags = "@disabled.by.default", strict = true, extraGlue = "com.algorand.algosdk.cucumber.shared")
+@CucumberOptions(publish = false, plugin = {"progress"}, tags = "@disabled.by.default", strict = true, extraGlue = "com.algorand.algosdk.cucumber.shared")
 public class RunCucumberUnitTest {
 }